pyproject.toml
[project]
name = "oh-my-brain"
version = "0.1.0"
description = "多Agent协作开发框架 - 让AI团队为你并行编程"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
authors = [{ name = "OH MY BRAIN Team" }]
keywords = ["agent", "multi-agent", "ai", "llm", "collaboration", "development"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyzmq>=25.0",
    "redis>=5.0",
    "httpx>=0.25",
    "tiktoken>=0.5",
    "gitpython>=3.1",
    "pyyaml>=6.0",
    "rich>=13.0",
    "typer>=0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.1",
    "mypy>=1.8",
    "pre-commit>=3.6",
]
docs = [
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.24",
]
dashboard = [
    "fastapi>=0.109",
    "uvicorn>=0.27",
    "websockets>=12.0",
]

[project.scripts]
oh-my-brain = "oh_my_brain.cli:app"
omb = "oh_my_brain.cli:app"

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/oh-my-brain"
Documentation = "https://YOUR_USERNAME.github.io/oh-my-brain"
Repository = "https://github.com/YOUR_USERNAME/oh-my-brain"
Issues = "https://github.com/YOUR_USERNAME/oh-my-brain/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/oh_my_brain"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/docs",
]

# ============ 工具配置 ============

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
]
ignore = ["E501"]  # line too long (handled by formatter)

[tool.ruff.lint.isort]
known-first-party = ["oh_my_brain"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["redis.*", "zmq.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src/oh_my_brain"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
